Issue#SC-1109 modified searchDtoMapper and resolve pr comments
