+++ user-org-actor/src/main/java/org/sunbird/user/UserReadActor.java
@@ -4,19 +4,12 @@ import org.sunbird.BaseActor;
 import org.sunbird.DaoImplType;
 import org.sunbird.actor.core.ActorConfig;
 import org.sunbird.actorOperation.UserActorOperations;
-import org.sunbird.common.factory.EsClientFactory;
-import org.sunbird.es.service.ElasticSearchService;
 import org.sunbird.exception.BaseException;
-import org.sunbird.helper.ElasticSearchHelper;
 import org.sunbird.request.Request;
 import org.sunbird.response.Response;
 import org.sunbird.user.dao.IUserESDao;
-import org.sunbird.user.dao.IUserOSDao;
 import org.sunbird.user.dao.UserDaoFactory;
 import org.sunbird.util.jsonkey.JsonKey;
-import scala.concurrent.Future;
-
-import java.util.Map;
 
 
 /**
@@ -36,7 +29,7 @@ public class UserReadActor extends BaseActor {
 
     @Override
     public void onReceive(Request request) throws Throwable {
-        if (request.getOperation().equalsIgnoreCase(UserActorOperations.READ_USER_BY_ID.getOperation())) {
+        if (UserActorOperations.READ_USER_BY_ID.getOperation().equalsIgnoreCase(request.getOperation())) {
             readUserById(request);
         } else {
             onReceiveUnsupportedMessage(this.getClass().getName());
@@ -44,8 +37,8 @@ public class UserReadActor extends BaseActor {
     }
 
     public void readUserById(Request request) throws BaseException {
-        Response response=userESDao.getUserById((String) request.getRequest().get(JsonKey.USER_ID));
-        sender().tell(response,self());
+        Response response = userESDao.getUserById((String) request.getRequest().get(JsonKey.USER_ID));
+        sender().tell(response, self());
     }
 
 }
